Update DeepSeek-V2-MoE ReadMe and tensor parallel mla support

+++ megatron_patch/model/qwen1_5/moe/router.py
@@ -160,7 +160,7 @@ class TopKRouter(Router):
         #top_logits, indices = torch.topk(logits, k=self.topk, dim=1)
         #scores = torch.softmax(top_logits, dim=-1, dtype=torch.float32).type_as(logits)
 
-        routing_weights = torch.softmax(logits, dim=1, dtype=torch.float)
+        routing_weights = torch.softmax(logits, dim=1, dtype=torch.float32).type_as(logits)
         scores, indices = torch.topk(routing_weights, k=self.topk, dim=-1)
 
         # Apply load balancing loss
